Fixed weird aliasing issue

+++ include/raycast.h
@@ -9,7 +9,7 @@
 void drawRays(SDL_Renderer *gRenderer, Player *player);
 
 //Handles 3D projection onto screen
-void draw3D(SDL_Renderer *gRenderer, Player *player, float rayAngle, int distanceTotal, int horizontalMapText, float shading, float rayX, float rayY,int i);
+void draw3D(SDL_Renderer *gRenderer, Player *player, float rayAngle, float distanceTotal, int horizontalMapText, float shading, float rayX, float rayY,int i);
 
 //Renders vertical Rays 
 float renderVerticalRays(Player *player, float rayAngle, float *verticalX, float *verticalY, float *rayX, float *rayY, int *verticalMapText);
